requirements.txt
# Core Framework
Flask==2.3.3
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0
python-socketio==5.9.0
eventlet==0.33.3

# Deep Learning
tensorflow==2.15.0
keras==2.13.1
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Computer Vision
opencv-python==4.8.0.76
opencv-contrib-python==4.8.0.76
dlib==19.24.1
face-recognition==1.3.0
mtcnn==0.1.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.13
pydub==0.25.1
scipy==1.11.2

# ML & Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
imbalanced-learn==0.11.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
matplotlib==3.7.2
seaborn==0.12.2

# Monitoring & Logging
python-json-logger==2.0.7